backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
python-socketio==5.11.0
phone-agent @ git+https://github.com/zai-org/Open-AutoGLM.git
